refactor(pages): keep the locale prefix helpers free of node-only imports
