+++ pages/src/queries/findPageByPath.ts
@@ -7,8 +7,9 @@ import type { PageCollectionConfig } from '../types/PageCollectionConfig.js'
 import type { PagesPluginConfig } from '../types/PagesPluginConfig.js'
 import type { FindPageByPathArgs, PageDocument, PageDocumentResult } from './types.js'
 
-import { parseLocalizedPath, resolveLocaleRouting } from '../utils/localeRouting.js'
+import { parseLocalizedPath } from '../utils/localePrefix.js'
 import { isPageCollectionConfig } from '../utils/pageCollectionConfigHelpers.js'
+import { resolveLocaleRouting } from '../utils/resolveLocaleRouting.js'
 import { ROOT_PAGE_SLUG } from '../utils/setRootPageVirtualFields.js'
 import { livenessConditions } from './liveness.js'
 import { buildPathCacheKey, type PathCacheEntry } from './pathCache.js'
