+++ pages/src/utils/localePrefix.ts
@@ -1,79 +1,11 @@
-import type { Payload, PayloadRequest } from 'payload'
-
-import { AsyncLocalStorage } from 'node:async_hooks'
+/**
+ * Pure locale-prefix helpers, shared by the server hooks and by the admin client components. The
+ * client bundle cannot resolve node built-ins, so this module must stay free of them — the routing
+ * resolver, which needs `node:async_hooks`, lives in `resolveLocaleRouting.ts` instead.
+ */
 
 import type { Locale } from '../types/Locale.js'
-import type { LocaleRouting, PagesPluginConfig } from '../types/PagesPluginConfig.js'
-
-/**
- * Marks the asynchronous scope a `localeRouting` resolver runs in. Everything the resolver awaits
- * inherits the mark, which is what separates a read performed *by* the resolver from a path
- * computation running concurrently *next to* it — the former must not wait for the resolver's
- * result, the latter must.
- */
-const resolvingRouting = new AsyncLocalStorage<true>()
-
-/**
- * The evaluation in flight or completed for one request. Keyed by the request object rather than
- * stored on `req.context`, which Payload replaces with a shallow copy on every nested local API
- * call — including the ones a resolver makes, which would fork the cache entry away mid-flight.
- */
-const routingByRequest = new WeakMap<PayloadRequest, Promise<LocaleRouting | undefined>>()
-
-/**
- * The locale routing in effect for the given request.
- *
- * A function resolver is evaluated once per request and its (possibly still pending) result is
- * shared, so computing the paths of fifty documents costs one evaluation.
- *
- * A resolver which itself reads a page collection would otherwise wait for its own result: the
- * read's `beforeRead` hook computes paths, which asks for the routing that is still being
- * resolved. Such reads are answered with the default routing (every locale prefixed) instead, so
- * paths seen inside a resolver carry no routing of their own.
- */
-export async function resolveLocaleRouting({
-  payload,
-  pluginConfig,
-  req,
-}: {
-  payload: Payload
-  pluginConfig: PagesPluginConfig | undefined
-  req: PayloadRequest | undefined
-}): Promise<LocaleRouting | undefined> {
-  const option = pluginConfig?.localeRouting
-  const localization = payload.config.localization
-
-  if (!option || !localization) {
-    return undefined
-  }
-
-  if (typeof option !== 'function') {
-    return validateRouting(option, localization.localeCodes)
-  }
-
-  if (!req) {
-    throw new Error(
-      'Resolving a page by path requires `req` when the plugin is configured with a `localeRouting` function, so the routing can be evaluated against the request.',
-    )
-  }
-
-  if (resolvingRouting.getStore()) {
-    return undefined
-  }
-
-  const cached = routingByRequest.get(req)
-
-  if (cached) {
-    return cached
-  }
-
-  const pending = resolvingRouting
-    .run(true, async () => await option({ req }))
-    .then((routing) => (routing ? validateRouting(routing, localization.localeCodes) : undefined))
-  routingByRequest.set(req, pending)
-
-  return pending
-}
+import type { LocaleRouting } from '../types/PagesPluginConfig.js'
 
 /** Whether the locale's paths carry a `/<locale>` prefix under the given routing. */
 export function isPrefixedLocale(
@@ -173,14 +105,4 @@ export function localeRoutingCacheToken(routing: LocaleRouting | undefined): str
     return '-'
   }
   return `${routing.primaryLocale}${routing.prefixPrimaryLocale === false ? '!' : ''}`
-}
-
-/** A typo in `primaryLocale` would silently rewrite every path, so it is rejected loudly. */
-function validateRouting(routing: LocaleRouting, localeCodes: Locale[]): LocaleRouting {
-  if (!localeCodes.includes(routing.primaryLocale)) {
-    throw new Error(
-      `[Pages Plugin] \`localeRouting.primaryLocale\` is "${routing.primaryLocale}", which is not one of the configured locales (${localeCodes.join(', ')}).`,
-    )
-  }
-  return routing
 }
